// blockchain/ergo/explorerApi/api.ts
export type AdditionalRegisters = Partial<Record<RegisterType, RegisterDetail>>;
export interface AssetInstanceInfo {
    /**
     * Token ID
     * @type {string}
     * @memberof AssetInstanceInfo
     */
    'tokenId': string;
    /**
     * Index of the asset in an output
     * @type {number}
     * @memberof AssetInstanceInfo
     */
    'index': number;
    /**
     * Amount of tokens
     * @type {number}
     * @memberof AssetInstanceInfo
     */
    'amount': bigint;
    /**
     * Name of a token
     * @type {string}
     * @memberof AssetInstanceInfo
     */
    'name'?: string;
    /**
     * Number of decimal places
     * @type {number}
     * @memberof AssetInstanceInfo
     */
    'decimals'?: number;
    /**
     * Type of a token (token standard)
     * @type {string}
     * @memberof AssetInstanceInfo
     */
    'type'?: string;
}
// ...
export interface DataInputInfo {
    /**
     * ID of the corresponding box
     * @type {string}
     * @memberof DataInputInfo
     */
    'boxId': string;
    /**
     * Number of nanoErgs in the corresponding box
     * @type {number}
     * @memberof DataInputInfo
     */
    'value': number;
    /**
     * Index of the input in a transaction
     * @type {number}
     * @memberof DataInputInfo
     */
    'index': number;
    /**
     * Modifier ID
     * @type {string}
     * @memberof DataInputInfo
     */
    'outputBlockId': string;
    /**
     * ID of the transaction outputting corresponding box
     * @type {string}
     * @memberof DataInputInfo
     */
    'outputTransactionId': string;
    /**
     * Index of the output corresponding this input
     * @type {number}
     * @memberof DataInputInfo
     */
    'outputIndex': number;
    /**
     * Hex-encoded string
     * @type {string}
     * @memberof DataInputInfo
     */
    'ergoTree': string;
    /**
     * Decoded address of the corresponding box holder
     * @type {string}
     * @memberof DataInputInfo
     */
    'address': string;
    /**
     *
     * @type {Array<AssetInstanceInfo>}
     * @memberof DataInputInfo
     */
    'assets'?: Array<AssetInstanceInfo>;
    /**
     *
     * @type {{ [key: string]: string; }}
     * @memberof DataInputInfo
     */
    'additionalRegisters': { [key: string]: string; };
}
// ...
export interface InputInfo {
    /**
     * ID of the corresponding box
     * @type {string}
     * @memberof InputInfo
     */
    'boxId': string;
    /**
     * Number of nanoErgs in the corresponding box
     * @type {number}
     * @memberof InputInfo
     */
    'value': number;
    /**
     * Index of the input in a transaction
     * @type {number}
     * @memberof InputInfo
     */
    'index': number;
    /**
     * Hex-encoded serialized sigma proof
     * @type {string}
     * @memberof InputInfo
     */
    'spendingProof'?: string;
    /**
     * Modifier ID
     * @type {string}
     * @memberof InputInfo
     */
    'outputBlockId': string;
    /**
     * ID of the transaction outputting corresponding box
     * @type {string}
     * @memberof InputInfo
     */
    'outputTransactionId': string;
    /**
     * Index of the output corresponding this input
     * @type {number}
     * @memberof InputInfo
     */
    'outputIndex': number;
    /**
     * Global index of the output corresponding this input
     * @type {number}
     * @memberof InputInfo
     */
    'outputGlobalIndex': number;
    /**
     * Height the output corresponding this input was created at
     * @type {number}
     * @memberof InputInfo
     */
    'outputCreatedAt': number;
    /**
     * Height the output corresponding this input was settled at
     * @type {number}
     * @memberof InputInfo
     */
    'outputSettledAt': number;
    /**
     * Hex-encoded string
     * @type {string}
     * @memberof InputInfo
     */
    'ergoTree': string;
    /**
     *
     * @type {string}
     * @memberof InputInfo
     */
    'ergoTreeConstants': string;
    /**
     *
     * @type {string}
     * @memberof InputInfo
     */
    'ergoTreeScript': string;
    /**
     * Decoded address of the corresponding box holder
     * @type {string}
     * @memberof InputInfo
     */
    'address': string;
    /**
     *
     * @type {Array<AssetInstanceInfo>}
     * @memberof InputInfo
     */
    'assets'?: Array<AssetInstanceInfo>;
    /**
     *
     * @type {{ [key: string]: string; }}
     * @memberof InputInfo
     */
    'additionalRegisters': { [key: string]: string; };
}
// ...
export interface OutputInfo {
    /**
     * Id of the box
     * @type {string}
     * @memberof OutputInfo
     */
    'boxId': string;
    /**
     * Id of the transaction that created the box
     * @type {string}
     * @memberof OutputInfo
     */
    'transactionId': string;
    /**
     * Id of the block a box included in
     * @type {string}
     * @memberof OutputInfo
     */
    'blockId': string;
    /**
     * Value of the box in nanoERG
     * @type {number}
     * @memberof OutputInfo
     */
    'value': bigint;
    /**
     * Index of the output in a transaction
     * @type {number}
     * @memberof OutputInfo
     */
    'index': number;
    /**
     * Global index of the output in the blockchain
     * @type {number}
     * @memberof OutputInfo
     */
    'globalIndex': number;
    /**
     * Height at which the box was created
     * @type {number}
     * @memberof OutputInfo
     */
    'creationHeight': number;
    /**
     * Height at which the box got fixed in blockchain
     * @type {number}
     * @memberof OutputInfo
     */
    'settlementHeight': number;
    /**
     * Serialized ergo tree
     * @type {string}
     * @memberof OutputInfo
     */
    'ergoTree': string;
    /**
     *
     * @type {string}
     * @memberof OutputInfo
     */
    'ergoTreeConstants': string;
    /**
     *
     * @type {string}
     * @memberof OutputInfo
     */
    'ergoTreeScript': string;
    /**
     * An address derived from ergo tree
     * @type {string}
     * @memberof OutputInfo
     */
    'address': string;
    /**
     *
     * @type {Array<AssetInstanceInfo>}
     * @memberof OutputInfo
     */
    'assets'?: Array<AssetInstanceInfo>;
    /**
     *
     * @type {{ [key: string]: string; }}
     * @memberof OutputInfo
     */
    'additionalRegisters': AdditionalRegisters;
    /**
     * Id of the transaction this output was spent by
     * @type {string}
     * @memberof OutputInfo
     */
    'spentTransactionId'?: string;
    /**
     *
     * @type {boolean}
     * @memberof OutputInfo
     */
    'mainChain': boolean;
}
// ...
export interface TransactionInfo {
    /**
     * Transaction ID
     * @type {string}
     * @memberof TransactionInfo
     */
    'id': string;
    /**
     * ID of the corresponding header
     * @type {string}
     * @memberof TransactionInfo
     */
    'blockId': string;
    /**
     * Height of the block the transaction was included in
     * @type {number}
     * @memberof TransactionInfo
     */
    'inclusionHeight': number;
    /**
     * Timestamp the transaction got into the network
     * @type {number}
     * @memberof TransactionInfo
     */
    'timestamp': number;
    /**
     * Index of a transaction inside a block
     * @type {number}
     * @memberof TransactionInfo
     */
    'index': number;
    /**
     * global index of a transaction in the blockchain
     * @type {number}
     * @memberof TransactionInfo
     */
    'globalIndex': number;
    /**
     * Number of transaction confirmations
     * @type {number}
     * @memberof TransactionInfo
     */
    'numConfirmations': number;
    /**
     *
     * @type {Array<InputInfo>}
     * @memberof TransactionInfo
     */
    'inputs'?: Array<InputInfo>;
    /**
     *
     * @type {Array<DataInputInfo>}
     * @memberof TransactionInfo
     */
    'dataInputs'?: Array<DataInputInfo>;
    /**
     *
     * @type {Array<OutputInfo>}
     * @memberof TransactionInfo
     */
    'outputs'?: Array<OutputInfo>;
    /**
     * Transaction size in bytes
     * @type {number}
     * @memberof TransactionInfo
     */
    'size': number;
}